Hardcode SSL as the encapsulated protocol to make UDP sessions work.
The Zeek-side DPD signatures for SSL require a TCP connection. That
check unfortunately always matches against the outer connection,
meaning it will fail for OpenVPN-over-UDP even though we're internally
using the TCP version of DPD for the inner layer. This would require a
Zeek change to fix / do differently (or a 2nd set of SSL signatures).